Refactor Services build error and resized container sizes for navbar and header
