+++ src/app/components/Home/Services/Services.module.css
@@ -1,25 +1,8 @@
 .title {
-  font-size: rem(34px);
   font-weight: 900;
-
-  @media (max-width: $mantine-breakpoint-sm) {
-    font-size: rem(24px);
-  }
-}
-
-.description {
-  max-width: rem(600px);
-  margin: auto;
-
-  &::after {
-    content: "";
-    display: block;
-    background-color: var(--mantine-color-blue-filled);
-    width: rem(45px);
-    height: rem(2px);
-    margin-top: var(--mantine-spacing-sm);
-    margin-left: auto;
-    margin-right: auto;
+  font-size: rem(32px);
+  @media (max-width: $mantine-breakpoint-xs) {
+    font-size: rem(28px);
   }
 }
 
@@ -32,7 +15,7 @@
   &::after {
     content: "";
     display: block;
-    background-color: var(--mantine-color-blue-filled);
+    background-color: var(--mantine-color-red-filled);
     width: rem(45px);
     height: rem(2px);
     margin-top: var(--mantine-spacing-sm);
