+++ src/app/components/Navbar/Navbar.module.css
@@ -1,6 +1,6 @@
 .header {
   height: 6em;
-  margin-bottom: 8px;
+  margin-bottom: rem(120px, 16px);
   background-color: var(--mantine-color-body);
 }
 
@@ -15,26 +15,10 @@
   display: flex;
   flex-direction: column;
   line-height: 1;
-  padding: 8px 12px;
+  padding: rem(8px) rem(12px);
   border-radius: var(--mantine-radius-sm);
   text-decoration: none;
   color: light-dark(var(--mantine-color-gray-7), var(--mantine-color-dark-0));
-  font-size: var(--mantine-font-size-sm);
+  font-size: var(--mantine-font-size-md);
   font-weight: 500;
 }
-[data-mantine-color-scheme] .link[data-active] {
-  background-color: var(--mantine-color-blue-filled);
-  color: var(--mantine-color-white);
-}
-.link:hover {
-  background-color: light-dark(var(--mantine-color-gray-0), var(--mantine-color-dark-6));
-}
-
-.dropdown {
-  display: block;
-}
-@media (min-width: 36em) {
-  .dropdown {
-    display: none;
-  }
-}/*# sourceMappingURL=Navbar.module.css.map */
