--- src/app/components/Navbar/Navbar.module.scss
@@ -1,46 +1,0 @@
-.header {
-    height: 6em;
-    margin-bottom: rem(120px, 16px);
-    background-color: var(--mantine-color-body);
-}
-
-.inner {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-/*link is for <a> in items*/
-.link {
-    display: flex;
-    flex-direction: column;
-    line-height: 1;
-    padding: rem(8px, 16px) rem(12px, 16px);
-    border-radius: var(--mantine-radius-sm);
-    text-decoration: none;
-    color: light-dark(var(--mantine-color-gray-7), var(--mantine-color-dark-0));
-    font-size: var(--mantine-font-size-sm);
-    font-weight: 500;
-
-    @mixin hover {
-        background-color: light-dark(var(--mantine-color-gray-0), var(--mantine-color-dark-6));
-    }
-
-    [data-mantine-color-scheme] &[data-active] {
-        background-color: var(--mantine-color-blue-filled);
-        color: var(--mantine-color-white);
-    }
-
-    &:hover {
-        background-color: light-dark(var(--mantine-color-gray-0), var(--mantine-color-dark-6));
-    }
-}
-
-//for mobile nav - burger menu
-.dropdown {
-    display: block;
-
-    @media(min-width: 36em) {
-        display: none;
-    }
-}
